requirements.txt
aiohttp==3.9.3
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0